extract shared Ray cluster and sidecar helpers into cookbook modules
The slime_disagg and miles_disagg helpers duplicated ~200 lines of
identical Ray cluster bring-up (start_ray_head, start_ray_worker,
get_modal_cluster_context, training_nodes) and sidecar process
management (start_sglang_sidecar, wait_http, terminate_process).

- New cookbook/ray_cluster.py: Ray head/worker startup, cluster context
  resolution, and diagnostic log dumping.
- New cookbook/sidecar_process.py: sidecar subprocess launch (parameterized
  on sidecar_module), HTTP liveness polling, and graceful process
  termination.
- Both helpers modules re-export the shared functions so existing callers
  (modal_train.py) continue to work without import changes.
- Trainer-specific code (prepare_config, build_train_cmd, smoke_flash_pool)
  stays in each trainer's helpers since it references trainer-specific
  config attributes and has genuinely different smoke-check strategies.
